Adjust total_kinetic_enery calculation
Let's assume that one needs to put energy into the system to change its
dynamical state; in this case, the kinetic energy in the system.
The change of the kinetic energy over time is its first difference. If
we assume that both to reduce and to increase the kinetic energy, one
needs to put energy into the system, then the sum of the absolute
changes in kinetic energy should correspond to the energy needed by the
system to function.
To a degree, in any case.